@Aurelienvd 2025-08-07
@xavieralt The PR fixing the issue has been r+ https://github.com/odoo/odoo/pull/222080